requirements.txt
beautifulsoup4==4.13.4
datasets==3.2.0
deepeval==2.5.2
detoxify==0.5.2
Flask==3.1.1
Flask_Cors==5.0.0
langchain==0.3.25
langchain_community==0.3.24
langchain_core==0.3.61
langchain_groq==0.3.2
langchain_openai==0.3.18
MarkupSafe==3.0.2
matplotlib==3.10.3
openai==1.82.0
pandas==2.2.3
pymongo==4.13.0
python-dotenv==1.1.0
qdrant_client==1.14.2
ragas==0.2.0
Requests==2.32.3
schedule==1.2.2
scikit_learn==1.6.1
sentence_transformers==3.3.1
